converts last settlement API method to use new request structure
