+++ web/src/types.ts
@@ -1,0 +1,8 @@
+export type Settlement = {
+  id: string;
+  name: string;
+  limit: number;
+  departing: number;
+  cc: number;
+  year: number;
+};
